# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceKernel API",
    "version": "1.0.0",
    "description": "The voice infrastructure layer for regulated enterprise. Create agents, place calls, ground answers in your knowledge, and stream every event into your own systems."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "List agents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Bypass the local snapshot and read through to the voice provider."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of agents."
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create an agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              },
              "example": {
                "name": "Card disputes",
                "systemPrompt": "You are a card disputes specialist for a retail bank. Answer only from the supplied policy documents. If unsure, escalate to a human.",
                "firstMessage": "Thanks for calling - how can I help with your card today?",
                "model": {
                  "provider": "anthropic",
                  "model": "claude-sonnet-5",
                  "temperature": 0.3
                },
                "voice": {
                  "provider": "11labs",
                  "voiceId": "matilda"
                },
                "transcriber": {
                  "provider": "deepgram",
                  "model": "nova-3"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created agent."
          }
        }
      }
    },
    "/v1/agents/{id}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Retrieve an agent",
        "responses": {
          "200": {
            "description": "The agent."
          }
        }
      },
      "patch": {
        "tags": [
          "Agents"
        ],
        "summary": "Update an agent",
        "responses": {
          "200": {
            "description": "The updated agent."
          }
        }
      },
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Delete an agent",
        "responses": {
          "200": {
            "description": "Deleted."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/agents/{id}/model": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Swap the agent model, preserving its prompt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "provider": "anthropic",
                "model": "claude-sonnet-5",
                "temperature": 0.3
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated agent."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/agents/{id}/prompt": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Update the agent prompt only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "systemPrompt": "You are…"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated agent."
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "vk_live_…"
      }
    },
    "schemas": {
...
      "AgentInput": {
        "type": "object",
        "required": [
          "name",
          "systemPrompt"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "systemPrompt": {
            "type": "string",
            "description": "The agent's instructions."
          },
          "firstMessage": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "anthropic",
                  "anthropic-bedrock",
                  "anyscale",
                  "cerebras",
                  "custom-llm",
                  "deep-seek",
                  "deepinfra",
                  "google",
                  "groq",
                  "inflection-ai",
                  "minimax",
                  "openai",
                  "openrouter",
                  "perplexity-ai",
                  "together-ai",
                  "voicekernel",
                  "xai"
                ]
              },
              "model": {
                "type": "string"
              },
              "temperature": {
                "type": "number"
              },
              "maxTokens": {
                "type": "integer"
              },
              "toolIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "knowledgeBaseId": {
                "type": "string"
              }
            }
          },
          "voice": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "11labs",
                  "azure",
                  "cartesia",
                  "custom-voice",
                  "deepgram",
                  "hume",
                  "inworld",
                  "lmnt",
                  "microsoft",
                  "minimax",
                  "neuphonic",
                  "openai",
                  "playht",
                  "rime-ai",
                  "sesame",
                  "smallest-ai",
                  "tavus",
                  "voicekernel",
                  "wellsaid",
                  "xai"
                ]
              },
              "voiceId": {
                "type": "string"
              }
            }
          },
          "transcriber": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "11labs",
                  "assembly-ai",
                  "azure",
                  "cartesia",
                  "custom-transcriber",
                  "deepgram",
                  "gladia",
                  "google",
                  "openai",
                  "soniox",
                  "speechmatics",
                  "talkscriber",
                  "voicekernel",
                  "xai"
                ]
              },
              "model": {
                "type": "string"
              },
              "language": {
                "type": "string"
              }
            }
          },
          "server": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              },
              "secret": {
                "type": "string"
              }
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "provider": {
            "type": "object",
            "additionalProperties": true,
            "description": "Escape hatch merged over the generated provider assistants."
          }
        }
      }
    }
  }
}